# internal/http/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Transactions Service API
  description: API for creating accounts and recording card transactions.
  version: 1.0.0
servers:
  - url: /
tags:
  - name: Health
  - name: Accounts
  - name: Transactions
paths:
  /health:
    get:
      tags: [Health]
      summary: Check service health
      operationId: getHealth
      responses:
        "200":
          description: The service is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /accounts:
    post:
      tags: [Accounts]
      summary: Create an account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
            example:
              document_number: "12345678901"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Retrieve an account
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          description: Account identifier.
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Account found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /transactions:
    post:
      tags: [Transactions]
      summary: Create a transaction
      description: Purchases and withdrawals require negative amounts. Credit vouchers require positive amounts.
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
            example:
              account_id: 1
              operation_type_id: 4
              amount: 123.45
      responses:
        "201":
          description: Transaction created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    CreateAccountRequest:
      type: object
      required: [document_number]
      properties:
        document_number:
          type: string
          example: "12345678901"
    Account:
      type: object
      required: [account_id, document_number]
      properties:
        account_id:
          type: integer
          format: int64
          example: 1
        document_number:
          type: string
          example: "12345678901"
    CreateTransactionRequest:
      type: object
      required: [account_id, operation_type_id, amount]
      properties:
        account_id:
          type: integer
          format: int64
          minimum: 1
          example: 1
        operation_type_id:
          type: integer
          format: int64
          description: "1: Normal Purchase, 2: Purchase with installments, 3: Withdrawal, 4: Credit Voucher."
          enum: [1, 2, 3, 4]
          example: 4
        amount:
          type: number
          format: double
          description: Negative for purchases and withdrawals; positive for credit vouchers.
          example: 123.45
    Transaction:
      type: object
      required: [transaction_id, account_id, operation_type_id, amount, event_date]
      properties:
        transaction_id:
          type: integer
          format: int64
          example: 1
        account_id:
          type: integer
          format: int64
          example: 1
        operation_type_id:
          type: integer
          format: int64
          enum: [1, 2, 3, 4]
          example: 4
        amount:
          type: number
          format: double
          example: 123.45
        event_date:
          type: string
          format: date-time
          example: "2026-06-01T12:00:00Z"
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: resource not found